refactor: rename user utils
