+++ src/use-cases/user/read-all-users/read-all-users-use-case.ts
@@ -1,6 +1,6 @@
 import { IUserDataQuery, IUserRepository, IUserVisibleData } from '@/use-cases/user/interfaces';
 import { IUseCase } from '@/use-cases/interfaces';
-import { getUserVisibleData } from '@/use-cases/user/util';
+import { getUserVisibleData } from '@/use-cases/user/utils';
 
 type Response = IUserVisibleData[] | [];
