--- src/use-cases/user/util/index.ts
@@ -1,1 +1,0 @@
-export * from '@/use-cases/user/util/user';
