+++ test/use-cases/user/delete-user/delete-user-use-case.spec.ts
@@ -2,7 +2,7 @@ import { IUserRepository } from '@/use-cases/user/interfaces';
 import { NonExistingUserError } from '@/use-cases/user/errors';
 import { UserBuilder } from '@/test/builders/user-builder';
 import { makeUserRepository } from '@/test/stubs';
-import { getUserVisibleData } from '@/use-cases/user/util';
+import { getUserVisibleData } from '@/use-cases/user/utils';
 import { DeleteUserUseCase } from '@/use-cases/user/delete-user';
 
 type SutTypes = {
@@ -39,6 +39,7 @@ describe('DeleteUserUseCase', () => {
       createdAt: new Date(),
       updatedAt: new Date(),
       isDeleted: false,
+      roles: [],
     };
     jest.spyOn(userRepository, 'findById').mockResolvedValue(findByIdReturn);
 
@@ -55,6 +56,7 @@ describe('DeleteUserUseCase', () => {
       createdAt: new Date(),
       updatedAt: new Date(),
       isDeleted: false,
+      roles: [],
     };
     jest.spyOn(userRepository, 'findById').mockResolvedValue(findByIdReturn);
     jest.spyOn(userRepository, 'deleteById').mockResolvedValue(findByIdReturn);
